+++ src/main/java/org/millie/www/MillieServer/book/Book.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.millie.www.MillieServer.user.UserBook;
+import org.millie.www.MillieServer.user.domain.UserBook;
 
 import java.util.ArrayList;
 import java.util.List;
