+++ src/main/java/org/millie/www/MillieServer/user/api/MemberContoller.java
@@ -1,6 +1,7 @@
-package org.millie.www.MillieServer.book;
+package org.millie.www.MillieServer.user.api;
 
 import lombok.RequiredArgsConstructor;
+import org.millie.www.MillieServer.book.BookService;
 import org.millie.www.MillieServer.book.dto.response.PostListGetResponse;
 import org.millie.www.MillieServer.common.dto.ApiResponse;
 import org.millie.www.MillieServer.common.httpmessage.SuccessMessage;
