+++ src/main/java/org/millie/www/MillieServer/user/repository/UserJpaRepository.java
@@ -1,5 +1,6 @@
-package org.millie.www.MillieServer.user;
+package org.millie.www.MillieServer.user.repository;
 
+import org.millie.www.MillieServer.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserJpaRepository extends JpaRepository<User, Long> {
